+++ gulpfile.babel.js
@@ -3,24 +3,34 @@
 import gulp from 'gulp';
 
 import abimis from './tasks/abimis';
-import nodeEnv from './tasks/node-env';
+import minimist from 'minimist';
+import { nodeDevEnv, nodeDistEnv } from './tasks/node-env';
 import clean from './tasks/clean';
 import html from './tasks/html';
 import { sass, sassCheck, sassDist } from './tasks/sass';
-import { jsLibs, jsLibsDist } from './tasks/js-libs';
-import { jsApp, jsAppDist } from './tasks/js-app';
+import { javascript, javascriptDist } from './tasks/javascript';
 import { assets, images, imagesDist } from './tasks/assets';
 import { humans, robots } from './tasks/txt';
 import server from './tasks/server';
+import serverRedux from './tasks/server-redux';
 import favicons from './tasks/favicons';
 import get from './tasks/get';
 
-import { HTML, SASS, ASSETS } from './settings';
+const PROJECT = minimist(process.argv.slice(2)).project;
+if (PROJECT) {
+    process.env.PROJECT = PROJECT;
+} else {
+    process.env.PROJECT = 'settings';
+}
+
+const HTML = require('./' + process.env.PROJECT).HTML;
+const SASS = require('./' + process.env.PROJECT).SASS;
+const ASSETS = require('./' + process.env.PROJECT).ASSETS;
 
 // This BUILD task:
 // Builds all the website files except the project's JS
 gulp.task('build',
-    gulp.series(clean, gulp.parallel(html, sass, jsLibs, assets, images))
+    gulp.series(clean, gulp.parallel(html, sass, assets, images))
 );
 
 // This WATCH function:
@@ -30,6 +40,9 @@ function watch() {
     gulp.watch(SASS.sourcePath, sass);
     gulp.watch(ASSETS.sourcePath, assets);
     gulp.watch(ASSETS.imagesSourcePath, images);
+
+    gulp.watch('./abimis_components/**/*.hbs', html);
+    gulp.watch('./abimis_components/**/*.scss', sass);
 }
 
 // This DEFAULT task:
@@ -38,14 +51,14 @@ function watch() {
 // Builds the project's JS
 // Starts watching all changes
 gulp.task('default',
-    gulp.series(abimis, 'build', server, gulp.parallel(jsApp, watch))
+    gulp.series(abimis, nodeDevEnv, 'build', javascript, server, gulp.parallel(watch))
 );
 
 // This BUILDDIST task:
 // Calls the SASSCHECK task
 // Builds all the website files except the project's JS
 gulp.task('buildDist',
-    gulp.series(clean, sassCheck, gulp.parallel(html, sassDist, jsLibsDist, assets, imagesDist, humans, robots))
+    gulp.series(clean, sassCheck, gulp.parallel(html, sassDist, assets, imagesDist, humans, robots))
 );
 
 // This WATCHDIST function:
@@ -64,7 +77,13 @@ function watchDist() {
 // Builds the project's JS
 // Starts watching all changes
 gulp.task('distribution',
-    gulp.series(abimis, nodeEnv, 'buildDist', server, gulp.parallel(jsAppDist, watchDist))
+    gulp.series(abimis, nodeDistEnv, 'buildDist', javascriptDist, server, gulp.parallel(watchDist))
+);
+
+// This REDUX task:
+// Runs the Redux server
+gulp.task('redux',
+    gulp.series(serverRedux)
 );
 
 // This FAVICONS task: