+++ src/helpers/helpers.js
@@ -1,6 +1,3 @@
-module.exports = {
-    // This helper is an example and has to be deleted in distribution
-    myFirstHelper: function (options) {
-        return options.fn(this);
-    },
+module.exports = function (options) {
+    return options.fn(this);
 }