+++ tasks/clean.js
@@ -1,8 +1,8 @@
-import { HTML } from '../settings';
 import rimraf from 'rimraf';
 
 // This CLEAN function:
-// Deletes the dist folder
+// Deletes the dest folder
 export default function clean(done) {
+    const HTML = require('../' + process.env.PROJECT).HTML;
     rimraf(HTML.destPath, done);
 }