+++ tasks/error-log.js
@@ -1,5 +1,0 @@
-// Building the general error management
-export default function (error) {
-    console.log(error);
-    this.emit('end');
-}