+++ tasks/html.js
@@ -1,19 +1,20 @@
 import gulp from 'gulp';
-import { HTML } from '../settings';
-import compileHandlebars from 'gulp-hb';
-import errorHandlebars from './error-handlebars';
+import iungo from 'iungo';
+import errorNotify from './error-notify';
 import browser from 'browser-sync';
 
 // This HTML function:
 // Compiles the Handlebars files including them in their respective files
 export default function html() {
+    const HTML = require('../' + process.env.PROJECT).HTML;
+
     return gulp.src(HTML.pages)
-        .pipe(compileHandlebars({
-            partials: HTML.partials,
+        .pipe(iungo({
             data: HTML.data,
+            partials: HTML.partials,
             helpers: HTML.helpers
         }))
-        .on('error', errorHandlebars)
+        .on('error', errorNotify)
         .pipe(gulp.dest(HTML.destPath))
         .pipe(browser.reload({ stream: true }));
 }