+++ tasks/node-env.js
@@ -1,6 +1,13 @@
-// This NODEENV function:
+// This NODEDEVENV function:
+// Sets NODE_ENV = 'development'
+export function nodeDevEnv(done) {
+    process.env.NODE_ENV = 'development';
+    done();
+}
+
+// This NODEDISTENV function:
 // Sets NODE_ENV = 'production'
-export default function nodeEnv(done) {
+export function nodeDistEnv(done) {
     process.env.NODE_ENV = 'production';
     done();
 }