+++ tasks/server.js
@@ -1,14 +1,16 @@
-import { HTML, PORT } from '../settings';
 import browser from 'browser-sync';
-import historyApiFallback from 'connect-history-api-fallback';
+import historyAPIFallback from 'connect-history-api-fallback';
 
 // This SERVER function:
 // Starts the server with live reload
 export default function server(done) {
+    const HTML = require('../' + process.env.PROJECT).HTML;
+    const PORT = require('../' + process.env.PROJECT).PORT;
+
     browser.init({
         server: {
             baseDir: HTML.destPath,
-            middleware: [historyApiFallback()] // Fixes issues with the HTML5 History API
+            middleware: [historyAPIFallback()] // Fixes issues with the HTML5 History API
         },
         port: PORT
     });